// src/app/shared/models/tax-invoices/tax-invoice-tab-line.ts
export interface TaxInvoiceTabLine {
    rowNumber: number //"TAB1_A1", "1"
    productName: string // "TAB1_A13", "електрична енергія"/"природний газ"/"Компенсація за недотримання гарантованих стандартів якості надання послуг ОСР"
    productCode: string // "TAB1_A131", "електрична енергія" - "2716000000" / "Компенсація за недотримання гарантованих стандартів якості надання послуг ОСР" - "35.13"
    unit: string // "TAB1_A14", "кВт·год", "м3", "грн" / одиниці вимір
    unitCode: string // "TAB1_A141",  "електрична енергія" - "0415" / компенсація - "2454" / м3 - ?
    quantity: number // "TAB1_A15" / споживання/кількість
    price: number //"TAB1_A16" "Ціна постачання одиниці товару\послуги" (tariff)
    liabilitiesAmount: number //"TAB1_A10" ToString("0.00").Replace(',', '.') "Обсяги постачання (база оподаткування) без урахування податку на додану вартість" (База опадаткування)
    tax: number // "TAB1_A20", .ToString("0.000000").Replace(',', '.'), "1") 6-digit after coma!!!!; / Сумма ПДВ
}